refactor components into separate sections, write a simple test for page
